+++ index.html
@@ -22,7 +22,7 @@
     <script>
       if (/*@cc_on!@*/ false || (!!window.MSInputMethodContext && !!document.documentMode))
         window.location.href =
-          'https://support.dmeng.net/upgrade-your-browser.html?referrer=' +
+          "https://support.dmeng.net/upgrade-your-browser.html?referrer=" +
           encodeURIComponent(window.location.href);
     </script>
   </head>
