+++ package.json
@@ -21,6 +21,7 @@
     "pinia-plugin-persistedstate": "^3.0.0",
     "swiper": "^9.3.2",
     "vue": "^3.3.4",
+    "vue-router": "4",
     "vue3-aplayer": "^1.7.3"
   },
   "devDependencies": {
