+++ src/main.js
@@ -3,6 +3,7 @@ import "@/style/style.scss";
 import App from "@/App.vue";
 // 引入 pinia
 import { createPinia } from "pinia";
+import router from "./router/index.js";
 import piniaPluginPersistedstate from "pinia-plugin-persistedstate";
 
 const app = createApp(App);
@@ -10,6 +11,7 @@ const pinia = createPinia();
 pinia.use(piniaPluginPersistedstate);
 
 app.use(pinia);
+app.use(router);
 app.mount("#app");
 
 // PWA
